+++ src/platform_impl/web/event_loop/runner.rs
@@ -1,9 +1,15 @@
-use super::{super::ScaleChangeArgs, backend, state::State};
+#[cfg(feature = "css-size")]
+use super::resize::ResizeState;
+use super::{backend, state::State};
+use crate::dpi::PhysicalSize;
 use crate::event::{Event, StartCause};
 use crate::event_loop::ControlFlow;
 use crate::window::WindowId;
 
 use instant::{Duration, Instant};
+use std::cell::Cell;
+#[cfg(feature = "css-size")]
+use std::cell::Ref;
 use std::{
     cell::RefCell,
     clone::Clone,
@@ -12,6 +18,7 @@ use std::{
     ops::Deref,
     rc::{Rc, Weak},
 };
+use web_sys::HtmlCanvasElement;
 
 pub struct Shared<T: 'static>(Rc<Execution<T>>);
 
@@ -27,8 +34,13 @@ pub struct Execution<T: 'static> {
     id: RefCell<u32>,
     all_canvases: RefCell<Vec<(WindowId, Weak<RefCell<backend::Canvas>>)>>,
     redraw_pending: RefCell<HashSet<WindowId>>,
-    destroy_pending: RefCell<VecDeque<WindowId>>,
+    destroy_pending: RefCell<VecDeque<(WindowId, HtmlCanvasElement)>>,
+    /// This is initially `None`, because it requires a handle on the runner.
+    #[cfg(feature = "css-size")]
+    resize_state: RefCell<Option<ResizeState>>,
+    #[cfg(not(feature = "css-size"))]
     scale_change_detector: RefCell<Option<backend::ScaleChangeDetector>>,
+    last_scale: Cell<f64>,
     unload_event_handle: RefCell<Option<backend::UnloadEventHandle>>,
 }
 
@@ -97,27 +109,61 @@ impl<T: 'static> Runner<T> {
 
 impl<T: 'static> Shared<T> {
     pub fn new() -> Self {
-        Shared(Rc::new(Execution {
+        let this = Shared(Rc::new(Execution {
             runner: RefCell::new(RunnerEnum::Pending),
             events: RefCell::new(VecDeque::new()),
             id: RefCell::new(0),
             all_canvases: RefCell::new(Vec::new()),
             redraw_pending: RefCell::new(HashSet::new()),
             destroy_pending: RefCell::new(VecDeque::new()),
+            #[cfg(feature = "css-size")]
+            resize_state: RefCell::new(None),
+            #[cfg(not(feature = "css-size"))]
             scale_change_detector: RefCell::new(None),
+            last_scale: Cell::new(backend::scale_factor()),
             unload_event_handle: RefCell::new(None),
-        }))
+        }));
+
+        #[cfg(feature = "css-size")]
+        {
+            *this.0.resize_state.borrow_mut() = Some(ResizeState::new(this.clone()));
+        }
+        #[cfg(not(feature = "css-size"))]
+        {
+            let runner = this.clone();
+            *this.0.scale_change_detector.borrow_mut() =
+                Some(backend::ScaleChangeDetector::new(move || {
+                    runner.handle_scale_changed(false)
+                }))
+        }
+
+        this
     }
 
     pub fn add_canvas(&self, id: WindowId, canvas: &Rc<RefCell<backend::Canvas>>) {
+        if self.num_canvases() == 0 {
+            // If we've had no canvases, the `ResizeObserver`s won't have been observing anything,
+            // and so we might not have noticed some changes in scale factor.
+            // So, make sure it's up to date.
+            self.0.last_scale.set(backend::scale_factor());
+        }
+
         self.0
             .all_canvases
             .borrow_mut()
             .push((id, Rc::downgrade(canvas)));
+
+        #[cfg(feature = "css-size")]
+        {
+            let resize_state = Ref::map(self.0.resize_state.borrow(), |resize_state| {
+                resize_state.as_ref().unwrap()
+            });
+            resize_state.observe(canvas.borrow().raw());
+        }
     }
 
-    pub fn notify_destroy_window(&self, id: WindowId) {
-        self.0.destroy_pending.borrow_mut().push_back(id);
+    pub fn notify_destroy_window(&self, id: WindowId, canvas: HtmlCanvasElement) {
+        self.0.destroy_pending.borrow_mut().push_back((id, canvas));
     }
 
     // Set the event callback to use for the event loop runner
@@ -134,14 +180,6 @@ impl<T: 'static> Shared<T> {
         let close_instance = self.clone();
         *self.0.unload_event_handle.borrow_mut() =
             Some(backend::on_unload(move || close_instance.handle_unload()));
-    }
-
-    pub(crate) fn set_on_scale_change<F>(&self, handler: F)
-    where
-        F: 'static + FnMut(ScaleChangeArgs),
-    {
-        *self.0.scale_change_detector.borrow_mut() =
-            Some(backend::ScaleChangeDetector::new(handler));
     }
 
     // Generate a strictly increasing ID
@@ -244,7 +282,16 @@ impl<T: 'static> Shared<T> {
     // `run_until_cleared` and `handle_scale_changed`, somewhere between emitting
     // `NewEvents` and `MainEventsCleared`.
     fn process_destroy_pending_windows(&self, control: &mut ControlFlow) {
-        while let Some(id) = self.0.destroy_pending.borrow_mut().pop_front() {
+        #[cfg(feature = "css-size")]
+        let resize_state = Ref::map(self.0.resize_state.borrow(), |resize_state| {
+            resize_state.as_ref().unwrap()
+        });
+
+        // `canvas` isn't used when we aren't using `ResizeObserver`.
+        #[cfg_attr(not(feature = "css-size"), allow(unused_variables))]
+        while let Some((id, canvas)) = self.0.destroy_pending.borrow_mut().pop_front() {
+            #[cfg(feature = "css-size")]
+            resize_state.unobserve(&canvas);
             self.0
                 .all_canvases
                 .borrow_mut()
@@ -287,82 +334,174 @@ impl<T: 'static> Shared<T> {
         }
     }
 
-    pub fn handle_scale_changed(&self, old_scale: f64, new_scale: f64) {
-        // If there aren't any windows, then there is nothing to do here.
-        if self.0.all_canvases.borrow().is_empty() {
+    /// Handle a change in scale factor, without any other information.
+    ///
+    /// `initial` is whether this is coming from the initial `ScaleChangeDetector`, rather than the `ResizeObserver`.
+    pub fn handle_scale_changed(&self, initial: bool) {
+        let canvases = self.0.all_canvases.borrow();
+        if canvases.is_empty() {
+            self.0.last_scale.set(backend::scale_factor());
             return;
         }
 
-        let start_cause = match (self.0.runner.borrow().maybe_runner())
-            .unwrap_or_else(|| unreachable!("`scale_changed` should not happen without a runner"))
-            .maybe_start_cause()
-        {
-            Some(c) => c,
-            // If we're in the exit state, don't do event processing
-            None => return,
-        };
+        let old_scale = self.0.last_scale.get();
+        let new_scale = backend::scale_factor();
+        // Whether or not the `ResizeObserver` will run; i.e., whether any canvases' logical sizes have changed.
+        let mut observer_will_run = false;
+
+        let resizes = self
+            .0
+            .all_canvases
+            .borrow()
+            .iter()
+            .filter_map(|(id, canvas)| {
+                // If the canvas was destroyed, it's not going to have been resized, so just skip it.
+                let canvas = canvas.upgrade()?;
+                let canvas = canvas.borrow();
+
+                // If the canvas isn't in the DOM, we don't need to handle the scale factor change for it.
+                let content_size = backend::inner_size(canvas.raw())?;
+
+                if canvas.size() != content_size.to_physical(old_scale) {
+                    observer_will_run = true;
+                }
+
+                Some((
+                    *id,
+                    canvas.raw().clone(),
+                    content_size.to_physical(new_scale),
+                ))
+            })
+            .collect();
+
+        // Don't handle the resizes yet if the `ResizeObserver` is going to run still.
+        if !(initial && observer_will_run) {
+            self.handle_resizes(resizes);
+        }
+    }
+
+    /// Handle a set of window resizes, as well as a scale factor change if any.
+    pub fn handle_resizes(
+        &self,
+        mut resizes: Vec<(WindowId, HtmlCanvasElement, PhysicalSize<u32>)>,
+    ) {
+        // Don't send resize events for destroyed canvases.
+        resizes.retain(|&(id, ..)| {
+            !self
+                .0
+                .destroy_pending
+                .borrow()
+                .iter()
+                .any(|&(other_id, ..)| other_id == id)
+        });
+
+        let scale = backend::scale_factor();
+
+        let scale_changed = self.0.last_scale.replace(scale) != scale;
+
         let mut control = self.current_control_flow();
 
-        // Handle the start event and all other events in the queue.
-        self.handle_event(Event::NewEvents(start_cause), &mut control);
-
-        // It is possible for windows to be dropped before this point. We don't
-        // want to send `ScaleFactorChanged` for destroyed windows, so we process
-        // the destroy-pending windows here.
-        self.process_destroy_pending_windows(&mut control);
-
-        // Now handle the `ScaleFactorChanged` events.
-        for &(id, ref canvas) in &*self.0.all_canvases.borrow() {
-            let canvas = match canvas.upgrade() {
-                Some(rc) => rc.borrow().raw().clone(),
-                // This shouldn't happen, but just in case...
-                None => continue,
+        if scale_changed {
+            let start_cause = match (self.0.runner.borrow().maybe_runner())
+                .and_then(|runner| runner.maybe_start_cause())
+            {
+                Some(c) => c,
+                // If the runner's not initialized yet or we're in the exit state, don't do event processing.
+                None => return,
             };
-            // First, we send the `ScaleFactorChanged` event:
-            let current_size = crate::dpi::PhysicalSize {
-                width: canvas.width() as u32,
-                height: canvas.height() as u32,
-            };
-            let logical_size = current_size.to_logical::<f64>(old_scale);
-            let mut new_size = logical_size.to_physical(new_scale);
-            self.handle_single_event_sync(
-                Event::WindowEvent {
-                    window_id: id,
-                    event: crate::event::WindowEvent::ScaleFactorChanged {
-                        scale_factor: new_scale,
-                        new_inner_size: &mut new_size,
-                    },
-                },
-                &mut control,
+            // Handle the start event and all other events in the queue.
+            self.handle_event(Event::NewEvents(start_cause), &mut control);
+
+            // It is possible for windows to be dropped before this point. We don't
+            // want to send `ScaleFactorChanged` for destroyed windows, so we process
+            // the destroy-pending windows here.
+            self.process_destroy_pending_windows(&mut control);
+
+            // Even if their physical sizes haven't changed, at least one of the two sizes of all the canvases must have changed when a scale factor change occurs.
+            // So, issue a resize event for all the other canvases as well, using their existing sizes.
+            let canvases = self.0.all_canvases.borrow();
+            let resized_ids: Vec<_> = resizes.iter().map(|&(id, ..)| id).collect();
+
+            resizes.extend(
+                canvases
+                    .iter()
+                    .filter(|&(id, ..)| !resized_ids.contains(id))
+                    .map(|(id, canvas)| {
+                        let canvas = canvas.upgrade().unwrap();
+                        let canvas = canvas.borrow();
+                        (*id, canvas.raw().clone(), canvas.size())
+                    }),
             );
+        }
+
+        for (id, canvas, size) in resizes.iter() {
+            let mut mut_size = *size;
+
+            let mut should_fire_resize = true;
+
+            // First, send a `ScaleFactorChanged` event if applicable.
+            if scale_changed {
+                self.handle_single_event_sync(
+                    Event::WindowEvent {
+                        window_id: *id,
+                        event: crate::event::WindowEvent::ScaleFactorChanged {
+                            scale_factor: scale,
+                            new_inner_size: &mut mut_size,
+                        },
+                    },
+                    &mut control,
+                );
+
+                if mut_size != *size {
+                    // Treat a change through this route the same way as `set_inner_size`, setting the CSS `width` and `height`.
+                    backend::set_inner_size(canvas, mut_size.into());
+                }
+            } else if size.width == canvas.width() && size.height == canvas.height() {
+                // If the canvas is already the correct size, don't send any resize events.
+                // This should only really happen immediately after a window is created,
+                // if the initial size based on CSS was already correct.
+                should_fire_resize = false;
+            }
 
             // Then we resize the canvas to the new size and send a `Resized` event:
-            backend::set_canvas_size(&canvas, crate::dpi::Size::Physical(new_size));
-            self.handle_single_event_sync(
-                Event::WindowEvent {
-                    window_id: id,
-                    event: crate::event::WindowEvent::Resized(new_size),
-                },
-                &mut control,
-            );
+            canvas.set_width(mut_size.width);
+            canvas.set_height(mut_size.height);
+
+            if should_fire_resize {
+                let event = Event::WindowEvent {
+                    window_id: *id,
+                    event: crate::event::WindowEvent::Resized(mut_size),
+                };
+
+                if scale_changed {
+                    self.handle_single_event_sync(event, &mut control);
+                } else {
+                    self.send_event(event);
+
+                    self.request_redraw(*id);
+                }
+            }
         }
 
-        // Process the destroy-pending windows again.
-        self.process_destroy_pending_windows(&mut control);
-        self.handle_event(Event::MainEventsCleared, &mut control);
+        if scale_changed {
+            // Process the destroy-pending windows again.
+            self.process_destroy_pending_windows(&mut control);
+            self.handle_event(Event::MainEventsCleared, &mut control);
 
-        // Discard all the pending redraw as we shall just redraw all windows.
-        self.0.redraw_pending.borrow_mut().clear();
-        for &(window_id, _) in &*self.0.all_canvases.borrow() {
-            self.handle_event(Event::RedrawRequested(window_id), &mut control);
-        }
-        self.handle_event(Event::RedrawEventsCleared, &mut control);
+            // Discard all the pending redraw as we shall just redraw all windows.
+            self.0.redraw_pending.borrow_mut().clear();
+            for &(window_id, _) in &*self.0.all_canvases.borrow() {
+                self.handle_event(Event::RedrawRequested(window_id), &mut control);
+            }
 
-        self.apply_control_flow(control);
-        // If the event loop is closed, it has been closed this iteration and now the closing
-        // event should be emitted
-        if self.is_closed() {
-            self.handle_loop_destroyed(&mut control);
+            self.handle_event(Event::RedrawEventsCleared, &mut control);
+
+            self.apply_control_flow(control);
+            // If the event loop is closed, it has been closed this iteration and now the closing
+            // event should be emitted
+            if self.is_closed() {
+                self.handle_loop_destroyed(&mut control);
+            }
         }
     }
 
@@ -465,7 +604,10 @@ impl<T: 'static> Shared<T> {
     fn handle_loop_destroyed(&self, control: &mut ControlFlow) {
         self.handle_event(Event::LoopDestroyed, control);
         let all_canvases = std::mem::take(&mut *self.0.all_canvases.borrow_mut());
-        *self.0.scale_change_detector.borrow_mut() = None;
+        #[cfg(feature = "css-size")]
+        {
+            *self.0.resize_state.borrow_mut() = None;
+        }
         *self.0.unload_event_handle.borrow_mut() = None;
         // Dropping the `Runner` drops the event handler closure, which will in
         // turn drop all `Window`s moved into the closure.
@@ -511,5 +653,28 @@ impl<T: 'static> Shared<T> {
             RunnerEnum::Pending => ControlFlow::Poll,
             RunnerEnum::Destroyed => ControlFlow::Exit,
         }
+    }
+
+    /// Retruns the number of canvases controlled by this runner which aren't pending destruction.
+    pub fn num_canvases(&self) -> usize {
+        self.0
+            .all_canvases
+            .borrow()
+            .iter()
+            .filter(|&&(id, _)| {
+                !self
+                    .0
+                    .destroy_pending
+                    .borrow()
+                    .iter()
+                    .any(|&(other_id, ..)| other_id == id)
+            })
+            .count()
+    }
+
+    #[cfg(feature = "css-size")]
+    /// Whether the scale factor has changed since the last `ScaleFactorChanged` event.
+    pub fn scale_factor_changed(&self) -> bool {
+        self.0.last_scale.get() != backend::scale_factor()
     }
 }